requirements.txt
python-docx
docx2txt
sqlalchemy
flask
gunicorn
pandas
psycopg2
pymongo
dnspython
